requirement.txt
# pip install -r requirement.txt
django
requests
requests_oauthlib
python-social-auth[django]
mecab-python3
